requirements.txt
autograd==1.7.0
boto3==1.34.19
datapackage==1.15.4
imageio==2.33.1
lifelines==0.29.0
matplotlib==3.8.2
nltk==3.9.1
numpy==2.1.1
openai==1.45.0
pandas==2.2.2
pooch==1.8.1
protobuf==5.28.1
psycopg2_binary==2.9.9
pydantic==2.9.1
pydicom==3.0.0
pyspark==3.5.2
pytest==8.0.2
python-dotenv==1.0.1
scikit_learn==1.4.1.post1
scipy==1.14.1
scispacy==0.5.4
setuptools==72.1.0
six==1.16.0
skimage==0.0
spacy==3.8.0
SQLAlchemy==2.0.29
tensorflow==2.17.0
torch==2.2.1
tqdm==4.66.1
